+++ src/Guide.php
@@ -18,6 +18,7 @@ use craft\services\UserPermissions;
 use wbrowar\guide\services\GuideService as GuideServiceService;
 use wbrowar\guide\twigextensions\GuideTwigExtension;
 use wbrowar\guide\models\Settings;
+use wbrowar\guide\widgets\AdminsLog;
 use wbrowar\guide\widgets\GuideWidget as GuideWidgetWidget;
 use wbrowar\guide\assetbundles\guide\GuideAsset;
 use wbrowar\adminbar\events\AdminBarRenderEvent;
@@ -33,6 +34,7 @@ use craft\services\Plugins;
 use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterUrlRulesEvent;
 
+use wbrowar\guide\widgets\WelcomeWidget;
 use yii\base\Event;
 
 /**
@@ -114,25 +116,47 @@ class Guide extends Plugin
 
         Event::on( Plugins::class, Plugins::EVENT_BEFORE_SAVE_PLUGIN_SETTINGS, function(PluginEvent $event) {
             if ($event->plugin->getHandle() === 'guide') {
-                $settings = Guide::$plugin->settings;
-                $vars = $settings->customVars;
-                $customVars = [];
+                //$settings = $event->plugin->getSettings();
+                $params = Craft::$app->request->getBodyParams();
+                $settingsFromPost = $params['settings'] ?? null;
+                $newSettings = [];
 
                 // Validate each variable
-                for ($i=0; $i<count($vars); $i++) {
-                    if (($vars[$i]['varKey'] ?? false) && ($vars[$i]['varValue'] ?? false)) {
-                        if ($vars[$i]['varType'] === 'password') {
-                            // if the variable is a password, encrypt it
-                            $vars[$i]['varValue'] = StringHelper::encenc($vars[$i]['varValue']);
-                        }
+                $vars = $settingsFromPost['customVars'] ?? null;
+                $customVars = [];
+                if ($vars ?? false) {
+                    for ($i=0; $i<count($vars); $i++) {
+                        if (($vars[$i]['varKey'] ?? false) && ($vars[$i]['varValue'] ?? false)) {
+                            if ($vars[$i]['varType'] === 'password') {
+                                // if the variable is a password, encrypt it
+                                $vars[$i]['varValue'] = StringHelper::encenc($vars[$i]['varValue']);
+                            }
 
-                        // add variable to new customVars array
-                        $customVars[] = $vars[$i];
+                            // add variable to new customVars array
+                            $customVars[] = $vars[$i];
+                        }
                     }
+                    $newSettings['customVars'] = $customVars;
+                }
+
+                // Validate nav items
+                $nav = $settingsFromPost['guideNav'] ?? null;
+                $navCount = 0;
+
+                if ($nav ?? false) {
+                    for ($i=0; $i<count($nav); $i++) {
+                        if (!empty($nav[$i]['id'])) {
+                            $navCount++;
+                        }
+                    }
+                    $guideNav = $navCount > 0 ? $settingsFromPost['guideNav'] : [];
+                    $newSettings['guideNav'] = $guideNav;
                 }
 
                 // add all custom variables back in to plugin settings
-                Guide::$plugin->setSettings(['customVars' => $customVars]);
+                if (!empty($newSettings)) {
+                    Guide::$plugin->setSettings($newSettings);
+                }
             }
         });
 
@@ -163,20 +187,15 @@ class Guide extends Plugin
                 $view->registerCss($this->getSettings()->customCss);
             });
 
-            // Register our site routes
-            //Event::on(
-            //    UrlManager::class,
-            //    UrlManager::EVENT_REGISTER_SITE_URL_RULES,
-            //    function (RegisterUrlRulesEvent $event) {
-            //    }
-            //);
-
             // Register our CP routes
             Event::on(
                 UrlManager::class,
                 UrlManager::EVENT_REGISTER_CP_URL_RULES,
                 function (RegisterUrlRulesEvent $event) {
                     $settings = $this->getSettings();
+
+                    $userCanDelete = Guide::$plugin->guide->userCanDeleteUserGuides();
+                    $userCanEdit = Guide::$plugin->guide->userCanEditUserGuides();
 
                     // Check to see if user's template exists
                     Craft::$app->view->setTemplateMode(View::TEMPLATE_MODE_SITE);
@@ -188,17 +207,32 @@ class Guide extends Plugin
                         $guideTemplatePathAsString = null;
                     }
 
-                    // Register CP Section home templates
-                    $event->rules['guide/home'] = ['template' => 'guide/home', 'variables' => ['pageContent' => $guideTemplatePathAsString, 'settings' => $settings, 'title' => 'Guide']];
-                    $event->rules['guide/components'] = ['template' => 'guide/home', 'variables' => ['settings' => $settings, 'title' => 'Guide Components']];
-
-                    // Register other CP Section templates
+                    // Register CP Section templates
                     $guideNav = $settings->guideNav;
-                    foreach ($guideNav as $item) {
-                        $templatePath = Craft::$app->view->resolveTemplate($item['template']);
-                        $templateAsString = $this->_getTemplateFileAsString($templatePath);
-                        $event->rules['guide/page/' . $item['id']] = ['template' => 'guide/home', 'variables' => ['pageContent' => $templateAsString, 'settings' => $settings, 'title' => $item['title']]];
+                    foreach ($guideNav as $key => $item) {
+                        $navItemVars = ['pageContent' => '', 'settings' => $settings, 'templatePath' => $item['template'] ?? null, 'title' => $item['title'], 'userCanEdit' => $userCanEdit, 'userCanDelete' => $userCanDelete];
+
+                        if ($item['userGuideId'] ?? false) {
+                            $navItemVars['userGuideId'] = $item['userGuideId'];
+                        } else if ($item['template'] ?? false) {
+                            $templatePath = Craft::$app->view->resolveTemplate($item['template']);
+                            $navItemVars['pageContent'] = $this->_getTemplateFileAsString($templatePath);
+                            $navItemPageContent = $navItemVars['pageContent'];
+                        }
+
+                        $event->rules['guide/page/' . $item['id']] = ['template' => 'guide/home', 'variables' => $navItemVars];
+                        $guideNav[$key]['pageContent'] = $navItemPageContent ?? null;
                     }
+
+                    // Register CP Section home templates
+                    $event->rules['guide/home'] = ['template' => 'guide/home', 'variables' => ['pageContent' => $guideTemplatePathAsString, 'settings' => $settings, 'templatePath' => $settings->guideTemplatePath, 'title' => 'Guide', 'guideNav' => $guideNav, 'userCanEdit' => $userCanEdit]];
+                    $event->rules['guide/components'] = ['template' => 'guide/home', 'variables' => ['settings' => $settings, 'title' => 'Guide Components', 'guideNav' => $guideNav]];
+                    $event->rules['guide/new'] = ['template' => 'guide/edit', 'variables' => ['settings' => $settings, 'title' => 'New User Guide', 'guideNav' => $guideNav, 'userCanEdit' => $userCanEdit, 'userCanDelete' => $userCanDelete]];
+                    $event->rules['guide/new/<templatePath:(.*)>'] = ['template' => 'guide/edit', 'variables' => ['settings' => $settings, 'title' => 'New User Guide', 'guideNav' => $guideNav, 'userCanEdit' => $userCanEdit, 'userCanDelete' => $userCanDelete]];
+                    $event->rules['guide/edit/<guideId:\d{1,}>'] = ['template' => 'guide/edit', 'variables' => ['settings' => $settings, 'title' => 'Edit User Guide', 'guideNav' => $guideNav, 'userCanEdit' => $userCanEdit, 'userCanDelete' => $userCanDelete]];
+                    $event->rules['guide/delete/<guideId:\d{1,}>'] = ['template' => 'guide/delete', 'variables' => ['userCanEdit' => $userCanEdit, 'userCanDelete' => $userCanDelete]];
+                    $event->rules['guide/admins-log'] = ['template' => 'guide/admins_log_settings', 'variables' => ['settings' => $settings, 'title' => 'Admin’s Log']];
+                    $event->rules['guide/welcome-widget'] = ['template' => 'guide/welcome_widget_settings', 'variables' => ['settings' => $settings, 'title' => 'Welcome Widget']];
 
                     Craft::$app->view->setTemplateMode(View::TEMPLATE_MODE_CP);
                 }
@@ -217,7 +251,9 @@ class Guide extends Plugin
                 Dashboard::class,
                 Dashboard::EVENT_REGISTER_WIDGET_TYPES,
                 function (RegisterComponentTypesEvent $event) {
+                    $event->types[] = AdminsLog::class;
                     $event->types[] = GuideWidgetWidget::class;
+                    $event->types[] = WelcomeWidget::class;
                 }
             );
 
@@ -238,6 +274,7 @@ class Guide extends Plugin
                 if ($id) {
                     $entry = Entry::find()
                         ->id($id)
+                        ->status(null)
                         ->one();
 
                     echo Guide::$plugin->guide->renderUserGuideTemplate('guide/_user_guide/user_guide_modal', $entry);
@@ -280,6 +317,19 @@ class Guide extends Plugin
         // Use the default name & icon, but customize the URL
         $navItem = parent::getCpNavItem();
         $navItem['url'] = 'guide/home';
+
+        $navItem['subnav'] = [
+            'home' => ['label' => 'Guide', 'url' => 'guide/home'],
+        ];
+
+        if (Craft::$app->getUser()->getIsAdmin() || Craft::$app->getUser()->checkPermission('editGuides')) {
+            $navItem['subnav']['components'] = ['label' => 'Guide Components', 'url' => 'guide/components'];
+        }
+        if (Craft::$app->getUser()->getIsAdmin()) {
+            $navItem['subnav']['adminsLogSettings'] = ['label' => 'Admin’s Log', 'url' => 'guide/admins-log'];
+            $navItem['subnav']['welcomeWidgetSettings'] = ['label' => 'Welcome Widget', 'url' => 'guide/welcome-widget'];
+        }
+
         return $navItem;
     }
 
@@ -306,7 +356,9 @@ class Guide extends Plugin
     {
         $settings = $this->getSettings();
 
-        $settings['customVars'] = Guide::$plugin->guide->prepareCustomVarsForSave($settings);
+        $newSettings = Guide::$plugin->guide->prepareCustomVarsForSave($settings);
+        $settings['customVars'] = $newSettings['customVars'];
+        $settings['guideNav'] = $newSettings['guideNav'];
 
         return Craft::$app->view->renderTemplate(
             'guide/settings',