+++ src/controllers/UserGuideController.php
@@ -10,6 +10,7 @@
 
 namespace wbrowar\guide\controllers;
 
+use craft\helpers\DateTimeHelper;
 use wbrowar\guide\Guide;
 
 use Craft;
@@ -88,6 +89,38 @@ class UserGuideController extends Controller
      *
      * @return mixed
      */
+    public function actionDeleteCpUserGuide()
+    {
+        $this->requirePostRequest();
+
+        $params = Craft::$app->getRequest()->getBodyParams();
+
+        if ($params['id'] ?? false) {
+            $userGuideId = $params['id'];
+            Guide::$plugin->guide->deleteUserGuide(['id' => $userGuideId]);
+
+            // update Guide CP nav
+            $nav = Guide::$plugin->getSettings()->guideNav;
+            $navCount = count($nav);
+
+            for ($i=0; $i<$navCount; $i++) {
+                if (isset($nav[$i]['userGuideId']) && ($nav[$i]['userGuideId'] == $userGuideId)) {
+                    unset($nav[$i]);
+                }
+            }
+
+            Guide::$plugin->guide->updateGuideCpNav($nav, true);
+
+            return $this->redirectToPostedUrl(true, 'guide/home');
+        }
+    }
+
+    /**
+     * Handle a request going to our plugin's actionDoSomething URL,
+     * e.g.: actions/guide/user-guide-controller/save-user-guide
+     *
+     * @return mixed
+     */
     public function actionGetUserGuideBody()
     {
         $this->requirePostRequest();
@@ -98,7 +131,9 @@ class UserGuideController extends Controller
         $sectionId = $params['sectionId'];
         $typeId = $params['typeId'];
 
-        $userGuide = Guide::$plugin->guide->getUserGuideForElementType($siteId, $sectionId, $typeId);
+        if (($siteId ?? false) && ($sectionId ?? false) && ($typeId ?? false)) {
+            $userGuide = Guide::$plugin->guide->getUserGuideForElementType($siteId, $sectionId, $typeId);
+        }
 
         if ($userGuide) {
             $variables = [
@@ -137,5 +172,139 @@ class UserGuideController extends Controller
         ]);
 
         return $this->asJson(Guide::$plugin->guide->saveUserGuide($userGuide));
+    }
+
+    /**
+     * Handle a request going to our plugin's actionDoSomething URL,
+     * e.g.: actions/guide/user-guide-controller/save-user-guide
+     *
+     * @return mixed
+     */
+    public function actionSaveCpUserGuide()
+    {
+        $this->requirePostRequest();
+
+        $params = Craft::$app->getRequest()->getBodyParams();
+
+        if (($params['content'] ?? false) &&
+            ($params['format'] ?? false) &&
+            ($params['title'] ?? false) &&
+            ($params['slug'] ?? false)) {
+
+            // save user guide
+            $userGuide = new UserGuide([
+                'authorId' => Craft::$app->getUser()->id,
+                'content' => $params['content'],
+                'elementType' => 'cp',
+                'format' => $params['format'],
+                'templatePath' => $params['templatePath'],
+            ]);
+
+            $userGuideId = Guide::$plugin->guide->saveUserGuide($userGuide, $params['id'] ?? null);
+
+
+            // update Guide CP nav
+            $nav = Guide::$plugin->getSettings()->guideNav;
+            $navCount = count($nav);
+
+            $userGuideInNav = false;
+            for ($i=0; $i<$navCount; $i++) {
+                if (isset($nav[$i]['userGuideId']) && $nav[$i]['userGuideId'] === $userGuideId) {
+                    $userGuideInNav = true;
+                    $nav[$i]['id'] = $params['slug'];
+                    $nav[$i]['title'] = $params['title'];
+                }
+            }
+
+            if (!$userGuideInNav) {
+                $nav[] = [
+                    'userGuideId' => $userGuideId,
+                    'id' => $params['slug'],
+                    'title' => $params['title'],
+                    'templatePath' => $params['templatePath'] ?? '',
+                    'permissions' => '',
+                    'admin' => '',
+                ];
+            }
+
+            Guide::$plugin->guide->updateGuideCpNav($nav, true);
+
+            return $this->redirectToPostedUrl(true, 'guide/page/' . $params['slug']);
+        }
+    }
+
+    /**
+     * Handle a request going to our plugin's actionDoSomething URL,
+     * e.g.: actions/guide/user-guide-controller/save-user-guide
+     *
+     * @return mixed
+     */
+    public function actionSaveWelcomeWidget()
+    {
+        $this->requirePostRequest();
+
+        $params = Craft::$app->getRequest()->getBodyParams();
+
+        // save welcome widget
+        $userGuide = new UserGuide([
+            'content' => $params['content'] ?? '',
+            'elementType' => 'welcomeWidget',
+            'format' => $params['format'],
+        ]);
+
+        Guide::$plugin->guide->saveUserGuide($userGuide);
+
+        // save new admin log
+
+        return $this->redirectToPostedUrl(true, 'guide/welcome-widget');
+    }
+
+    /**
+     * Handle a request going to our plugin's actionDoSomething URL,
+     * e.g.: actions/guide/user-guide-controller/save-user-guide
+     *
+     * @return mixed
+     */
+    public function actionUpdateAdminsLog()
+    {
+        $this->requirePostRequest();
+
+        $params = Craft::$app->getRequest()->getBodyParams();
+
+        // save new admin log entry
+        if ($params['content'] ?? false) {
+            $content = '';
+
+            if ($params['date'] ?? false) {
+                $logDate = date('M j, Y', strtotime($params['date']['date']));
+                $content .= '## ' . $logDate . "\n\n";
+            }
+
+            if ($params['version'] ?? false) {
+                $content .= '#### ' . $params['version'] . "\n\n";
+            }
+
+            $content .= $params['content'];
+
+            // save welcome widget
+            $userGuide = new UserGuide([
+                'content' => $content ?? '',
+                'elementType' => 'adminsLog',
+                'format' => 'markdown',
+            ]);
+
+            Guide::$plugin->guide->saveUserGuide($userGuide);
+        }
+
+        // delete selected admin log entriesc
+        if (($params['deleteLogEntries'] ?? false)) {
+            foreach($params['deleteLogEntries'] as $key => $value) {
+                if ($value) {
+                    Guide::$plugin->guide->deleteUserGuide(['id' => $key]);
+                }
+            }
+        }
+
+        return $this->redirectToPostedUrl(true, 'guide/admins-log');
     }
 }